Refactored tests into correct clases
