+++ _tests_/client.test.js
@@ -3,9 +3,10 @@ import Client from '../src/Client';
 import Message from '../src/Message';
 import Logger from '../src/Logger';
 
-describe ('raceResults service enhanced',() => {
+describe ('Client',() => {
     // SUT is Client   
     jest.mock("../src/Message");
+    jest.mock("../src/Client");
     jest.mock("../src/Logger");  // don't need real logger here so mock
 
     let client;
@@ -18,8 +19,15 @@ describe ('raceResults service enhanced',() => {
         client = new Client('Client1',category, logger);
         message = new Message(category);
     });
-     
-    // should be test of client or make raceresult interact with client
+
+    test('None Subscribed clients should not recieve message',() => {
+        const receiveMock =  jest.spyOn(client, "receive");
+
+        raceResultsService.send(message);
+        expect(receiveMock).not.toHaveBeenCalledWith(message);
+    });
+
+    
     test('Client subscribedCategory should be same as the message category', () => {
         raceResultsService.addSubscriber(client);
         raceResultsService.send(message);
@@ -46,46 +54,14 @@ describe ('raceResults service enhanced',() => {
         
         expect(receiveMock).not.toHaveReturnedWith(true);
     });
-});
 
-describe('Message sent by RaceResults Service Should log Date and text of each message',() => {
-    // SUT = Client
-    jest.mock("../src/Message");
-    jest.mock("../src/Logger");
-
-    const raceResultsService = new RaceResultsService(); 
-
-    const category = 'Boat Race';
-    const msgDate = "20/06/2022";
-    const msgText = "Hello from Logger";
-    let logger;
-    let client;
-    let message; 
-
-    beforeEach(() => {    
-        logger = new Logger();
-        client = new Client('Client1',category, logger);
-        message = new Message(category,msgText,msgDate);
-    });
-    
-    // test('Client logIncomingMessage function should be called with correct message ',() => {  
-    //     const logIncomingMessageMock = jest.spyOn(client, "logIncomingMessage"); // remember to set this up before client is used
-
-    //     raceResultsService.addSubscriber(client);
-    //     raceResultsService.send(message);
-        
-    //     expect(logIncomingMessageMock).toHaveBeenCalledWith(message);
-    // });
-
-    test('Logger LogMessage function Should be called with correct message',() => {   
-        const logMessageMock = jest.spyOn(logger, "logMessage")
-
+    test('removed subscriber does not receive message', () => {
+        const receiveMock =  jest.spyOn(client, "receive");
+            
         raceResultsService.addSubscriber(client);
+        raceResultsService.removeSubscriber();
         raceResultsService.send(message);
-        
-        expect(logMessageMock).toHaveBeenCalledWith(message);
+
+        expect(receiveMock).not.toHaveBeenCalledWith(message);
     });
 });
-
-
-// put these in client test no race results
