+++ _tests_/notes.txt
@@ -1,0 +1,10 @@
+Test client and logger seperately
+create client class test and logger tests
+word the test descriptions correctly
+
+Test send(message) in RaceResultsService
+
+Test Logger in its own test class.
+
+
+Test each function in each class
