+++ _tests_/raceResultsService.test.js
@@ -8,33 +8,9 @@ jest.mock("../src/Message");
 const raceResultsService = new RaceResultsService();
 
 describe('RaceResultsService',() => {   
-    test('Subscribed Client should receive Message', () => {
-        // SUT is raceResultsService
-        
-        const client = new Client();
-        const message = new Message();
-            
-        raceResultsService.addSubscriber(client);
-        raceResultsService.send(message);
-        expect(client.receive).toHaveBeenCalledWith(message);
-    });
+    test('addSubscriber',() =>{});
 
-    test('None Subscribed clients should not recieve message',() => {
-        const message = new Message();
-        const client = new Client();
+    test('removeSubscriber',() =>{});
 
-        raceResultsService.send(message);
-        expect(client.receive).not.toHaveBeenCalledWith(message);
-    });
-
-    test('removed subscriber does not receive message', () => {
-        const client = new Client();
-        const message = new Message();
-            
-        raceResultsService.addSubscriber(client);
-        raceResultsService.removeSubscriber();
-        raceResultsService.send(message);
-
-        expect(client.receive).not.toHaveBeenCalledWith(message);
-    });
+    test('send',() =>{});
 });
